+++ src/tm_devices/helpers/constants_and_dataclasses.py
@@ -1024,7 +1024,7 @@ USB_MODEL_ID_LOOKUP: Final[Mapping[SupportedModels, USBTMCConfiguration]] = Mapp
             vendor_id=TEKTRONIX_USBTMC_VENDOR_ID, model_id="0x0345"
         ),
         SupportedModels.AFG31K: USBTMCConfiguration(
-            vendor_id=TEKTRONIX_USBTMC_VENDOR_ID, model_id="0x035E"
+            vendor_id=TEKTRONIX_USBTMC_VENDOR_ID, model_id="0x0358"
         ),
         SupportedModels.SMU2450: USBTMCConfiguration(
             vendor_id=KEITHLEY_USBTMC_VENDOR_ID, model_id="0x2450"