+++ tests/test_config_parser.py
@@ -537,6 +537,11 @@ def test_remove_device() -> None:
             "USB0::0x0699::0x0345::C0123123::0::INSTR",
         ),
         (
+            "device_type=AFG,connection_type=USB,address=AFG31052-B0123123",
+            "AFG 1",
+            "USB0::0x0699::0x0358::B0123123::0::INSTR",
+        ),
+        (
             "device_type=SCOPE,address=3,connection_type=SERIAL",
             "SCOPE 1",
             "ASRL3::INSTR",